Move @JsonRPCApi to a proper package
Closes: #49
